requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==4.2
django-ckeditor-5==0.2.18
django-cors-headers==4.7.0
django-jazzmin==3.0.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==25.0
pillow==11.2.1
psycopg2==2.9.10
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
whitenoise==6.9.0
jinja2